updated the IMY processor file
decided to calculate the initial gravity vector before reading IMU data instead of assuming the IMU physical orientation. This will account for a dynamic gravitational filter based on the initial position of the component. This could potentially prevent the drastic changes in orientation through the accumulation of errors over time, especially when the car is stationary